Move optional package prompts to shared agent ownership
