+++ backend/src/flow44/ai/agents/execute/prompts.py
@@ -7,7 +7,7 @@ from typing import Any, Literal, overload
 
 from jinja2 import ChoiceLoader, Environment, FileSystemLoader, TemplateNotFound
 
-from flow44.ai.agents.execute.optional_packages import (
+from flow44.ai.agents.optional_packages import (
     OPTIONAL_PACKAGES,
     OptionalPackagePrompt,
     allowed_import_names,
