+++ backend/src/flow44/ai/agents/templates/_dependency_rules.jinja2
