+++ backend/src/flow44/ai/agents/templates/optional_packages/recharts/codegen_rules.jinja2
