+++ backend/src/flow44/ai/agents/templates/optional_packages/recharts/fix_errors_rules.jinja2
