+++ backend/src/flow44/ai/agents/templates/optional_packages/recharts/merge_rules.jinja2
